requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
DateTime==5.4
fastapi==0.110.0
idna==3.6
pydantic==2.6.3
pydantic_core==2.16.3
python-dotenv==1.0.1
pytz==2024.1
setuptools==69.1.1
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
typing_extensions==4.10.0
zope.interface==6.2